refactor(workflows): extract pure state model from canvas-editor (phase 1)
First, safest phase of the workflow-canvas-editor.tsx split — the pure
graph/state model, zero JSX.

lib/canvas-editor-state.ts holds: the layout constants (NODE_WIDTH/HEIGHT/
Y_SPACING/STANDARD_NODE_STYLE/TRIGGER+APPEND_NODE_ID/RESERVED_STEP_IDS), the
BUILTIN_STEP_LABELS map, the EditorState interface, the step→node-data /
id / body / clone builders (stepNodeData, nextStepId, defaultStepBody,
cloneStep), the legacy-layout heuristics, and the seed/derive/auto-arrange
functions (seedEdges, seedState, deriveNodes, autoArrangeState, samePosition,
sameMeasurement). All pure over EditorState + WorkflowDefinition + xyflow
types + lib/dagre-layout — now unit-testable without jsdom, following the
lib/dagre-layout / lib/node-renderer-registry precedent for client-only lib
modules.

The component imports them back and sheds the now-unused Edge / layoutNodes /
NodePosition imports. RESERVED_STEP_IDS (a test-asserted contract passed to
NodeConfigDrawer) and all other public shapes are preserved. 1,803 -> 1,564.

Pure behavior-preserving relocation — functions are byte-identical.

Verified: typecheck + eslint + workflow-canvas-editor.test (29/0, asserts the
selected-node className, insertable-edge data shape, RESERVED_STEP_IDS, and
layout.positions payload) + workflows suite (420/0) + full suite (5124/0) +
binary build + isolated-home boot smoke (Workflows loads; rebuilt client.js
→ 200).

The dead-renderer deletion (the Editor*Node renderers shadowed by the registry)
is behavior-touching and stays its own follow-up step, along with the
nodes/drawer/hook extractions and the appendix's redesigns.
